requirements.txt
h5py==3.13.0
joblib==1.5.1
numpy==1.24.3
pandas==1.5.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.3.0
scipy==1.10.1
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
